Question3
When we use implement Runnable, the order of the output will simultaneously be organized. Also, whenever the code moves from the first thread to the second thread, it will start at the first thread with the first value, then changed to the second thread corresponded with the first value and so on. If we wanted to keep it alike the output in the Question 1, we have to use a multi-Lock.